Improved mechanism that avoids running multiple instances of FreeSSM
Problem:
On UNIX-systems, QSharedMemory objects survive if the attached process is not detached on exit (e.g. on crashes, kills,...)

Solution:
If the SMO already exists when starting the application, we detach and then reattach it again to find out if the process that created it (the other application instance) is still alive.
If it is dead, QSharedMemory::detach() also detaches the dead process() and releases the SMO (=> reattachment fails).
If it is still running, the SMO is NOT released and reattachment succeeds.
This behavior of QSharedMemory is undocumented and we should be aware of changes in future Qt-versions.