Modified project file: fixed install target
Translation binaries (*.qm files) do not exist when calling qmake for the first time in a clean sourcecode directory.
As a result of that, qmake creates makefiles without including the translation binaries in the install target section.

The only known solution is to call qmake again after creating the translation binaries so that the makefiles are rebuilt.