Improved test mode state observation / DTC memory title determination
The test mode state is now read permanently during diagnostic codes reading (on ECUs).
It is emitted along with detected DTCs and the D-Check-state, so that the content widget can check it permanently and change the memory/table title if necessary.

This mainly fixes a (more or less theoretical) bug:
The DC-memory/-table title was not updated, if a "keen" user opened or closed test mode connector during the "Clear Memory"-procedure while ignition is switched off.