Changed Diagnostic Codes definition handling
With the recently added DTC-address block, we had lots of definitions for currently unused/unknown DCs.
This patch removes all definitions for unknown or unused DCs and add empty definitions for DCs that are to be ignored.
Handling of unused/unknown/ignored DCs is now done in SSMprotocol::evaluateDCdataByte():
- DCs without matching definitions are displayed as "Unknown" with a corresponding address information
- DCs with matching but empty definitions are ignored

The benefits are:
- reduced memory usage (RAM and binary file)
- new address blocks can be added easier
- "Features" information in the "About"-dialog now only includes the number of DCs that are completely defined (address AND code+title)