Added IDs for Powertrain Control Units (PCM)
Powertrain Control Unit = combined Engine + Transmission Control Units (PCM=ECU+TCU).
PCMs reply to both physical addresses (0x10 and 0x18) with the same IDs and flagbytes.
The system ID is always the same as for a corresponding ECU and they are only used together with E-4AT-transmissions.
So we will identify all CUs that reply to the TCU-physical address with a ECU-system ID as "E-4AT".
There is no overlapping with IDs of "real" TCMs.