Implemented workaround for a Qt-bug occuring under MS Windows
Under some (not yet completely explored) circumstances, multiple pressed()-signals
are emitted when QPushButtons are pressed while moving the mouse pointer.
This occures on MS Windows OSs only.
The simplest workaround is to connect all buttons to the released()-signal
instead (as it is still the better choice than the clicked()-signal).