serialCOM (Linux): fixed setting/getting of non-standard baudrates for drivers which only support the "new" method using the BOTHER baud constant
Some drivers (e.g. the ftdi_so-module for FTDI-USB-Serial-Convertors) only support the "new" BOTHER-method for non-standard baudrates, which currently is not supported by the GNU C-library.
BOTHER is defined as CBAUDEX | B0 which is equal to the baudrate bitmask. Unfortunately, the c_ispeed/c_ospeed-fields of the termios-struct are equal to c_cflag & CBAUD, too, so that no baudrate can be detected in this case.
The only solution is to use the Linux-ioctls directly instead of C-library functions.

This is not a nice situation, as it is a big step backwards from Unix/POSIX-compatibility and also makes the code more complicated.
Hopefully, the GNU C-libray will support this feature in future versions (I filed a bug-report at glibc-bugzilla).