serialCOM (Linux): fixed function SendBreak()
Since commit e9cecaa5dc79455f686810dc0f10457db6899ec0, serialCOM uses Linux-ioctls instead of C-library-functions.
The Linux-implementation of the TCSBRK-ioctl differs from the C-Lib-implemention (and also other systems). Although it can be used to send a 250ms-break, it is mostly designed to wait until a transmission i completed.
For break-sending, Linux defines the TCSBRKP-ioctl for compatibility. Unfortunately, its argument is interpreted a multiplicator of 100ms, so that the only way to send breaks with finer grained durations is to use the TIOCSBRK/TIOCCBRK-ioctls and to do the timing manulay with a usleep().

I noticed, that the GNU C-library-implementation of POSIX-function tcsendbreak() is also only capable of sending 100ms-step-breaks. So the SendBreak() function didn't work as expected in the past, too...