CU-contents: Fix calculation of min. nr. of table rows
After the latest layout changes, it turned out that the minimum nr. of table rows in CUcontent_Adjustments is not calculated correctly in some cases.
I did some investigations and this seems to be cause by a bug in Qt 4.5.3.
This commit fixes the calculation by switching back to the old calculation method using widget height and QHeaderview->viewport()->height().