ATcommandControlledDiagInterface: add experimental support for SSM2-over-ISO14230 with the ELM327
The ELM327 v1.4b is currently the only known interface which can - in theory - be configured to work with the SSM2-over-ISO14230-protocol.
Unfortunately the communication is extremely unstable. It turned out that the reason is the inter-byte timing value used by the ELM327.
ECUs seem to accept 0-5ms while ISO-14230 specifies 5-20ms. The ELM327 uses a value of ~5ms.

I've contacted ELMelectronics and maybe they will add a command for adjusting the timing values in future versions.
So let's keep this peace of code and disable it for now.
It can be enabled for experiments by defining __ENABLE_SSM2_ISO14230_EXPERIMENTAL_SUPPORT__.