serialCOM (Linux): improve compatibility with drivers for non-standard baud rates
Some devices/drivers (such as the pl2303 USB-serial converter) support the
ioctl TIOCGSERIAL, but not the TIOCSSERIAL ioctl.
Do not throw an error in this case and fall back to the BOTHER baud rate
encoding method.

Some drivers also might return an error for ioctl TIOCSSERIAL, if they do not
support the ASYNC_LOW_LATENCY flag.
In this case retry without this flag.

Also fix some typos and improve the debugging code a bit.