CUcontent_MBsSWs::startMBSWreading(): fix and simplify MB/SW data consistency validation
The data consisteny check currently succeeds if the list sizes are different,
which is a bug.
Add a == operator to class MBSWmetadata_dt, which allows us to check the equality of
two std::vector<MBSWmetadata_dt> directly using the == operator.
This includes the vector size comparison and hence fixes the bug.