Control Unit dialogs: disable mode/function selection buttons until a connection has been established
The current code calls setup() from the constructor, which makes the connection
establishing progress dialog appear immediately. Because this dialog is modal,
the mode/function selection buttons are currently protected until a connection
with the control unit has been established.

But in the future setup() will be made public and called separately, so we need
to explicitly protect the selection buttons by disabling them right at the
beginnning in the constructor and enabling them when setup() has finished.