Control Unit dialogs: refactor info widget handling
Create an abstract info widget and derive the 3 info widgets from it,
which includes unifiying function names for setting the system description.
This makes it possible to replace the individual
displaySystemDescriptionAndID(...) functions in the Control Unit dialogs
with a common version of this function in class ControlUnitDialog.

Rename functions fillInfoWidget(...) to displayExtendedCUinfo(...) to better
describe what they are doing.
Also pass the pointers to the SSMPdev and the info widget to this function
instead of accessing the global members.

Don't store pointers to the info widgets in the individual Control Unit
dialogs anymore, they are no longer needed.
Also make the info widget visible (by calling QWidget::show() in
ControlUnitDialog::setInfoWidget() insead of the constructors of the individual
Control Unit dialogs.