SSMP2communication_core: always clear Tx+Rx buffers before sending a new message
The interface buffers can contain incomplete messages, e.g. due to temporary
"ignition off" (which is done on purpose during the Clear Memory procedure).
This _should_ not happen with J2534- and AT-command controlled interfaces
(unless they are buggy), but it can't be avoided with serial pass-through
interfaces.
So better always flush the buffers before sending a new SSM2-protocol message.
There is no notable impact on the communication performance.