J2534DiagInterface::read(): do not try to detect the start of messages from control units
At least for SSM2 we can't assume that a message from the control unit
is always delivered with a single PASSTHRU_MSG.
The used timings exceed the limits defined in ISO-14230, so interfaces
can't always detect SSM2 message ends/starts via timeouts properly.

Apart from that, SAE-J2534 leaves much room for interpretation when it
comes to mapping incoming control unit messages to PASSTHRU_MSGSs.

Just write any received data to a buffer and return it to the caller.
It's up to him to assemble the control unit messages.