linux/serialCOM::SetPortSettings(...): fix custom baud rate encoding in case the driver reports 0 as baud_base
The driver may report 0 (or even a negative value) as baud_base in
struct serial_struct.

The reason for this is
    a) the driver is buggy
    b) the driver wants to tell us that the baud rate is not calculated
       and/or set using a baud rate base value and a divisor

In both cases, we need to encode the custom baud rate with the new
method (using c_ispeed + c_ospeed of struct termios2) or even use
standard baud rate encoding (using the next nearest standard baud rate).