SSMprotocol1::setupCUdata(...): do not call resetCUdata() if definitions are missing or invalid
Calling resetCUdata() resets _state to state_needSetup, which causes
getSysID() + getROMID() to fail.
We want to be able to retrieve/display both IDs even if defintions are not
available, so omit this call in this case.
resetCUdata() will then be called on the next call of setupCUdata(...)
or when the object is destructed.