Refactor and extend Diagnostic Codes handling
SSM1 and SSM2 ECUs report Diagnostic Codes using bit fields, which is already supported.

But SSM1 ECUs may use further methods, which assign the whole value
read from a designated DC address to a single DC.

The following assignment methods are known:
1) hexadecimal representation of the raw value is used as code
1) decimal representation of the raw value is used as code
3) raw value is assigned to DC using a list

ECUs may also use different DC lists, which differ in the amount of supported DCs
and the meaning of codes.

Refactor and extend the Diagnostic Codes handling to support all this.

NOTE: SSM1 definitions need to be updated/converted to the new XML format !