SSM1 definitions: improve XML format with regards to Diagnostic Codes
It turned out that direct scaling can not be used as often as desired.
Getting rid of it and using list scaling instead simplifes Diagnostic
Code definitions in the DATA_COMMON section and makes it much easier to
maintain them.
This comes at the cost of an increased complexity for directly scalable
addresses. Fortunately, only a few of them exist and the possible values
are limited.

NOTE: SSM1 definitions need to be updated/converted to the new XML format !