SSMDefinitionsInterface: introduce methods to retrieve data of special MBs/SWs
For special MBs/SWs such as "Engine Speed", "Ignition Switch", "Test
Mode" and "D-Check", hardcoded data (address, bit, scaling) is currently
used.
This was sufficient as long as we required them for ECUs using Ax10xx defintions
only.

Now support for actuator tests of SSM1 ECUs is going to be added and we need to
evaluate these MBs/SWs with them, too.
Hence, we add new methods to retrieve the ECU specific data of the special
MBs/SWs to the definition interface(s).