SSMprotocol: add method to find out if the ECU provides engine speed information
The engine speed informaton is used to find out if the engine is running
or not, which is a prerequise for e.g. actuator test start/stop.
Actuator test support for SSM1 ECUs has been added recently and many of them
do not support/use the engine speed information.
So the user needs to know whether the engine state can be retrieved with
isEngineRunning(...) or not.